fixed many compiler warnings and errors in x64 build in Visual Studio 2008
git-svn-id: file://localhost/home/tolik/svnrepo/trunk@1830 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ opencv/apps/haartraining/cvcommon.cpp
@@ -75,7 +75,7 @@ int icvMkDir( const char* filename )
     p = path;
     for( ; ; )
     {
-        pos = strcspn( p, "/\\" );
+        pos = (int)strcspn( p, "/\\" );
 
         if( pos == (int) strlen( p ) ) break;
         if( pos != 0 )
